// src/app/demo/api/wholeKey.ts
export interface wholeKey {
    AlertSent: number;
    Batch?: string;
    City?: string;
    EMail?: string;
    FirstName?: string;
    HouseNumber?: number;
    ID: string;
    LastName?: string;
    LoanPeriod?: number;
    Misc?: string;
    Name?: string;
    Number?: number;
    Street?: string;
    Timestamp?: string;
    kAlertSent: number;
    kID?: string;
    kLoanPeriod?: number;
    kName?: string;
    kTimestamp?: string;
    kMisc?: string;
}